Segregate the LLR and channel compensation related common functions into different files
