+++ openair1/PHY/nr_phy_common/CMakeLists.txt
@@ -1,6 +1,8 @@
 # SPDX-License-Identifier: LicenseRef-CSSL-1.0
 
-add_library(nr_phy_common src/nr_phy_common.c)
+add_library(nr_phy_common src/nr_phy_common.c
+                          src/nr_channel_compensation.c
+                          src/nr_compute_llr.c)
 target_link_libraries(nr_phy_common PRIVATE UTIL PHY_COMMON)
 target_include_directories(nr_phy_common PUBLIC inc/)
