+++ openair1/PHY/nr_phy_common/inc/nr_phy_common.h
@@ -377,93 +377,6 @@ int nr_get_ssb_start_sc(int scs,
                         int ssb_sco,
                         frequency_range_t freq_range);
 
-/**
- * @brief Common channel compensation function shared by DL (PDSCH) and UL (PUSCH) paths.
- *
- * Computes matched-filter output (rxComp) and channel magnitude arrays used for LLR
- * computation. MRC across Rx antennas is performed inline: for each layer, contributions
- * from all Rx antennas are accumulated into rxComp[layer * nb_rx_ant][symbol * buffer_length].
- * Uses AVX2 (256-bit SIMD) for throughput.
- *
- * @param buffer_length   Number of complex samples per symbol (must be a multiple of 8)
- * @param nb_rx_ant       Number of Rx antennas
- * @param nb_layers       Number of spatial layers
- * @param rxFext          Extracted received signal [nb_rx_ant][buffer_length]
- * @param chFext          Extracted channel estimates [nb_layers][nb_rx_ant][buffer_length]
- * @param ch_maga         Output magnitude array for threshold 'a' [nb_layers][buffer_length]
- * @param ch_magb         Output magnitude array for threshold 'b' [nb_layers][buffer_length]
- * @param ch_magc         Output magnitude array for threshold 'c' [nb_layers][buffer_length]
- * @param rxComp          Output compensated signal; row [l * nb_rx_ant] holds the MRC result
- *                        for layer l at offset [symbol * buffer_length]
- * @param rho             Tx-correlation matrix [nb_layers][nb_layers][buffer_length], or NULL
- * @param mod_order       Modulation order (2=QPSK, 4=16QAM, 6=64QAM, 8=256QAM)
- * @param symbol          OFDM symbol index (used to compute offset into rxComp rows)
- * @param output_shift    Right-shift applied after each complex multiply
- */
-void nr_channel_compensation(uint32_t buffer_length,
-                             int nb_rx_ant,
-                             int nb_layers,
-                             c16_t rxFext[nb_rx_ant][buffer_length],
-                             c16_t chFext[nb_layers][nb_rx_ant][buffer_length],
-                             c16_t ch_maga[nb_layers][buffer_length],
-                             c16_t ch_magb[nb_layers][buffer_length],
-                             c16_t ch_magc[nb_layers][buffer_length],
-                             c16_t **rxComp,
-                             c16_t (*rho)[nb_layers][buffer_length],
-                             int mod_order,
-                             uint32_t symbol,
-                             uint32_t output_shift);
-
-void nr_compute_llr(c16_t *rxdataF_comp,
-                    c16_t *ch_mag,
-                    c16_t *ch_magb,
-                    c16_t *ch_magc,
-                    int16_t *llr,
-                    uint32_t nb_re,
-                    uint8_t symbol,
-                    uint8_t mod_order);
-
-void nr_qpsk_llr_2layer(c16_t *stream0_in, c16_t *stream1_in, int16_t *stream0_out, c16_t *rho01, uint32_t length);
-
-void nr_qam16_llr_2layer(c16_t *stream0_in,
-                         c16_t *stream1_in,
-                         c16_t *ch_mag,
-                         c16_t *ch_mag_i,
-                         int16_t *stream0_out,
-                         c16_t *rho01,
-                         uint32_t length);
-
-void nr_qam64_llr_2layer(c16_t *stream0_in,
-                         c16_t *stream1_in,
-                         c16_t *ch_mag,
-                         c16_t *ch_mag_i,
-                         int16_t *stream0_out,
-                         c16_t *rho01,
-                         uint32_t length);
-
-void nr_compute_ML_llr(c16_t *rxdataF_comp0,
-                       c16_t *rxdataF_comp1,
-                       c16_t *ch_mag0,
-                       c16_t *ch_mag1,
-                       int16_t *llr_layers0,
-                       int16_t *llr_layers1,
-                       c16_t *rho0,
-                       c16_t *rho1,
-                       uint32_t nb_re,
-                       uint8_t mod_order);
-
-uint8_t nr_mmse_2layers(c16_t **rxdataF_comp,
-                        uint32_t buffer_length,
-                        int nb_rx_ant,
-                        int nb_layers,
-                        c16_t ch_mag[nb_layers][buffer_length],
-                        c16_t ch_magb[nb_layers][buffer_length],
-                        c16_t ch_magc[nb_layers][buffer_length],
-                        c16_t ul_ch_estimates_ext[][nb_rx_ant][buffer_length],
-                        unsigned short nb_rb,
-                        unsigned char mod_order,
-                        int shift,
-                        unsigned char symbol,
-                        int length,
-                        uint32_t noise_var);
+#include "nr_channel_compensation.h"
+#include "nr_compute_llr.h"
 #endif
